Add tick duration argument to testdriver Action
We want to allow the users to define the duration for every tick, so we
add an argument to Action class which will be a duration for all
ticks. If there is a tick which need a different duration, we can call
addTick or pause to pass a different value.

Bug: 606367
Change-Id: I7e5d88557f5f4a4c63dbf85985c30881aee9f739