Make table's clientWidth/Height include its entire border sizes.
Table with border-collapse: collapse
shares border between the container and its cells.
And clientWidth/Height returns the value which includes
the halved border sizes used for layout,
but it does not correspond to the definition of clientWidth/Height.
This CL adds the remaining halved border sizes
to clientWidth/Height when retrieving them.

As a result, when border-collapse: collapse
is specified in table, clientWidth/Height
and offsetWidth/Height are equal now, respectively.

Bug: 978019
Change-Id: I2693eaf38b882fe4a41f931ec1312ecf53352f27