[EventTiming] Replace onload tests
The onload-related tests were meant to test the performance timeline buffering
behavior. But now that Event Timing is not exposed on the timeline, those tests
don't really make much sense. Instead, a test is added to test that the
timestamps corresponding to two 'consecutive' clickAndBlockMain() calls behave
as one would expect.

Bug: 987236
Change-Id: Icc798fe1e0684a52992a83ba4de931481e1ded2a